[core cib] [projects default] [sys shell]
 * fixed some minor issues found by splint

+++ core/cib.c
@@ -7,14 +7,14 @@ void cib_init(cib_t *cib, unsigned int size) {
 }
 
 int cib_avail (cib_t *cib) {
-    return cib->write_count - cib->read_count;
+    return (int) (cib->write_count - cib->read_count);
 }
 
 int cib_get(cib_t *cib) {
     int avail = cib_avail (cib);
 
     if (avail > 0) {
-        return cib->read_count++ & ~cib->complement;
+        return (int) (cib->read_count++ & ~cib->complement);
     } 
 
     return -1;
@@ -24,7 +24,7 @@ int cib_put(cib_t *cib) {
     int avail = cib_avail (cib);
 
     if ((int)(avail + cib->complement) < 0 ) {
-        return cib->write_count++ & ~(cib->complement);
+        return (int) (cib->write_count++ & ~(cib->complement));
     }
 
     return -1;
